refactor: move all types to src/Types/<Page>/Types.ts and update all imports
