+++ 9.4.2/nextjs/src/Types/Repository/Types.ts
@@ -1,0 +1,11 @@
+export interface IRepository {
+  id: string
+  githubUrl: string
+  owner: string
+  name: string
+}
+
+export interface AddRepositoryModalProps {
+  open: boolean
+  onClose: () => void
+}
